requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
asttokens==2.2.1
    # via stack-data
backcall==0.2.0
    # via ipython
colorama==0.4.6
    # via ipython
decorator==5.1.1
    # via ipython
django==3.2.19
    # via
    #   -r requirements.in
    #   django-browser-reload
    #   django-currentuser
    #   django-phonenumber-field
    #   django-rich
    #   django-taggit
django-browser-reload==1.10.0
    # via -r requirements.in
django-currentuser==0.5.3
    # via -r requirements.in
django-phonenumber-field==7.1.0
    # via -r requirements.in
django-rich==1.6.0
    # via -r requirements.in
django-taggit==4.0.0
    # via -r requirements.in
executing==1.2.0
    # via stack-data
ipython==8.14.0
    # via -r requirements.in
jedi==0.18.2
    # via ipython
markdown-it-py==3.0.0
    # via rich
matplotlib-inline==0.1.6
    # via ipython
mdurl==0.1.2
    # via markdown-it-py
parso==0.8.3
    # via jedi
phonenumbers==8.13.15
    # via -r requirements.in
pickleshare==0.7.5
    # via ipython
pillow==9.5.0
    # via -r requirements.in
pip-lock==2.10.0
    # via -r requirements.in
prompt-toolkit==3.0.38
    # via ipython
pure-eval==0.2.2
    # via stack-data
pygments==2.15.1
    # via
    #   ipython
    #   rich
pytz==2023.3
    # via django
rich==13.4.2
    # via django-rich
six==1.16.0
    # via asttokens
sqlparse==0.4.4
    # via django
stack-data==0.6.2
    # via ipython
traitlets==5.9.0
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.6.3
    # via asgiref
wcwidth==0.2.6
    # via prompt-toolkit
